Implement FASTER correction for distance to reference channel
When the EEG data is references using a single channel, you will see
that channels that are further away from the reference will have signal
with larger amplitude. The outlier detection in FASTER needs to take
this into account. Note that this will only work for a single reference
electrode. If the data has an average reference of a mean reference,
this does not work.